backend/services/api-server/requirements.txt
email-validator==2.2.0
fastapi==0.115.12
google-cloud-secret-manager==2.23.3
httpx==0.27.2
pydantic==2.11.5
pydantic-settings==2.9.1
python-dotenv==1.1.0
python-jose==3.5.0
supabase==2.15.2
uvicorn==0.34.0
boto3==1.38.38
redis==6.2.0